+++ Modules/core/Model/CoreInstall.php
@@ -99,6 +99,12 @@ class CoreDB extends Model {
         Configuration::getLogger()->debug('set ac_anticorps counters, done!');
     }
 
+    public function repair371() {
+        Configuration::getLogger()->info("Run repair script 371 (PR #371)");
+        $usersInfo = new UsersInfo();
+        $usersInfo->addColumn("users_info", "organization", "varchar(255)", "");
+        Configuration::getLogger()->info("Run repair script 371 (PR #371)");
+    }
 
     public function upgrade_v0_v1() {